Fix genai tests
